fix: sticky KV cache bug causing 2x slowdown + drafter FP16 default
- TargetModel.verify(): remove permanent _kv_ok=False after successful
  full-forward fallback. A one-time PKV format mismatch should not
  disable KV cache for the entire prompt. Reset _kv_ok=True after
  successful fallback so next step retries incremental caching.
  This was the dominant cause of slowdown — target was doing O(L)
  full forwards every step instead of O(k) incremental forwards.
- DraftModel: change default dtype from float32 to float16.
  5x FP32 autoregressive forwards per step were eating ~50% of the
  step budget, negating speculative decoding speedup.
- Add --mlflow-uri CLI flag to point experiments at remote MLflow
  tracking server (e.g. http://host:5000/).
- Runner._setup_mlflow(): call mlflow.set_tracking_uri() when configured.

+++ src/core/models/draft_model.py
@@ -40,7 +40,7 @@ class DraftModel:
         self,
         model_name_or_path: str,
         device: str = "cuda",
-        dtype: torch.dtype = torch.float32,
+        dtype: torch.dtype = torch.float16,
         **model_kwargs,
     ) -> None:
         logger.info("Loading drafter tokenizer from %s", model_name_or_path)
